requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.32
openai==0.27.0
gitpython==3.1.27
requests==2.27.1
Werkzeug==2.0.3
